error handling

+++ src/task/task.controller.ts
@@ -1,6 +1,6 @@
 import { Controller } from '@nestjs/common';
 import { TaskService } from './task.service';
-import { GrpcMethod } from '@nestjs/microservices';
+import { GrpcMethod, RpcException } from '@nestjs/microservices';
 import {
   CreateTaskRequest,
   ListTasksRequest,
@@ -18,6 +18,17 @@ export class TaskController {
 
   @GrpcMethod('TaskService')
   createTask(data: CreateTaskRequest) {
+    if (
+      !data.task.title ||
+      !data.task.description ||
+      !data.task.dueDate ||
+      !data.task.status
+    ) {
+      throw new RpcException({
+        code: 400,
+        message: 'The task title, description, dueDate or Status is missing',
+      });
+    }
     const newTask = data.task;
 
     return this.taskService.create(newTask as any);
@@ -27,6 +38,13 @@ export class TaskController {
   async ListTasks(request: ListTasksRequest): Promise<ListTasksResponse> {
     const tasks = await this.taskService.findAll();
     console.log({ tasks });
+
+    if(!tasks) {
+      throw new RpcException({
+        code: 404,
+        message: 'No tasks were found',
+      });
+    }
 
     const res = ListTasksResponse.create({
       tasks: tasks.map((t) =>
@@ -43,6 +61,19 @@ export class TaskController {
 
   @GrpcMethod('TaskService')
   updateTask(data: UpdateTaskRequest) {
+    
+    if (
+      !data.task.title ||
+      !data.task.description ||
+      !data.task.dueDate ||
+      !data.task.status
+    ) {
+      throw new RpcException({
+        code: 400,
+        message: 'The task title, description, dueDate or Status is missing',
+      });
+    }
+
     const task = data.task;
 
     return this.taskService.update(task.id, task as any);
@@ -52,12 +83,26 @@ export class TaskController {
   deleteTask(data: DeleteTaskRequest) {
     const id = data.id;
 
+    if (!id) {
+      throw new RpcException({
+        code: 404,
+        message: 'The task id is missing or not found',
+      });
+    }
+
     return this.taskService.remove(id);
   }
   
   @GrpcMethod('TaskService')
   getTask(data: GetTaskRequest) {
     const id = data.id;
+
+    if (!id) {
+      throw new RpcException({
+        code: 404,
+        message: 'The task was not found',
+      });
+    }
 
     return this.taskService.remove(id);
   }
